modify architecture

+++ ultralytics/nn/modules/block.py
@@ -514,14 +514,14 @@ class C2f_Equalize(nn.Module):
         """Forward pass through C2f layer."""
         y = list(self.cv1(x).chunk(2, 1))
         y.extend(m(y[-1]) for m in self.m)
-        return self.eq(self.cv2(torch.cat(y, 1)))
+        return self.eq(self.cv2(torch.cat(y, 1)).astype(torch.uint8))
         
 
     def forward_split(self, x):
         """Forward pass using split() instead of chunk()."""
         y = list(self.cv1(x).split((self.c, self.c), 1))
         y.extend(m(y[-1]) for m in self.m)
-        return self.eq(self.cv2(torch.cat(y, 1)))
+        return self.eq(self.cv2(torch.cat(y, 1)).astype(torch.uint8))
     
 class C2f_Ghost_Conv(nn.Module):
     """Faster Implementation of CSP Bottleneck with 2 convolutions."""
